refactor reporting batch cancel route
