+++ src/app/router_registry.py
@@ -226,6 +226,7 @@ from app.routers.proposal_workflow_evidence import (
     router as proposal_workflow_evidence_router,
 )
 from app.routers.proposals import router as proposals_router
+from app.routers.reporting_batch_cancel import cancel_router as reporting_batch_cancel_router
 from app.routers.reporting_batch_controls import (
     controls_router as reporting_batch_controls_router,
 )
@@ -438,6 +439,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(reporting_batches_router)
     app.include_router(reporting_batch_status_router)
     app.include_router(reporting_batch_controls_router)
+    app.include_router(reporting_batch_cancel_router)
     app.include_router(reporting_batch_retry_router)
     app.include_router(reporting_batch_lease_recovery_router)
     app.include_router(reporting_batch_run_once_router)
